fix: updated docstrings to refer to float as np.float for consistency. Addressed by Abdul

+++ src/pystats_norm/rnorm.py
@@ -9,9 +9,9 @@ def rnorm(n, mean=0, sd=1):
     ----------
     n : int
         The number of random variables to be simulated.
-    mean : float, optional
+    mean : np.float, optional
         The mean value of the normal distribution. Default is 0.
-    sd : float, optional
+    sd : np.float, optional
         The standard deviation of the normal distribution. Default is 1.
     
     Returns
@@ -26,19 +26,19 @@ def rnorm(n, mean=0, sd=1):
     array([6.3245, 4.5983])
     """
 
-    #Checking for invalid inputs
+    # Checking for invalid inputs
     if not isinstance(n, int):
         raise ValueError('n must be an integer!')
     if n<0:
         raise ValueError('n must be a positive integer!')
-    if not isinstance(mean, (int, float)):
+    if not isinstance(mean, (int, np.float)):
         raise ValueError('The mean value must be a number!')
-    if not isinstance(sd, (int, float)):
+    if not isinstance(sd, (int, np.float)):
         raise ValueError('The standard deviation value must be a number!')
     if sd<0:
         raise ValueError('The standard deviation must be a positive number!')
     
-    #Return an empty array if n = 0 is passed
+    # Return an empty array if n = 0 is passed
     if n == 0:
         return np.array([])
